+++ GithubBackup.Cli.Tests/Commands/Github/Backup/BackupRunnerTests.cs
@@ -118,7 +118,8 @@ public class BackupRunnerTests
                     _ansiConsole.Profile.Out.Writer,
                     _ansiConsole.Profile.Out.Writer
                 )
-            )
+            ),
+            _ansiConsole
         );
     }
 }