+++ GithubBackup.Core.Tests/Github/Clients/GithubApiClientCacheTests.cs
@@ -5,6 +5,7 @@ using GithubBackup.Core.Github.Clients;
 using GithubBackup.Core.Github.Credentials;
 using GithubBackup.Core.Utils;
 using GithubBackup.TestUtils.Logging;
+using System.IO.Abstractions.TestingHelpers;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
@@ -28,6 +29,7 @@ public class GithubApiClientCacheTests
         tokenStore.GetAsync().Returns(Token);
 
         _sut = new GithubApiClient(
+            new MockFileSystem(),
             new MemoryCache(new MemoryCacheOptions()),
             tokenStore,
             new DateTimeOffsetProvider(),