+++ GithubBackup.Core/Github/Migrations/DownloadMigrationOptions.cs
@@ -7,7 +7,9 @@ public sealed class DownloadMigrationOptions(
     IDirectoryInfo destination,
     int? numberOfBackups = null,
     bool overwrite = true,
-    TimeSpan? medianFirstRetryDelay = null
+    TimeSpan? medianFirstRetryDelay = null,
+    Action<string>? onTemporaryFileCreated = null,
+    Action<long, long?>? onDownloadProgress = null
 )
 {
     public long Id { get; } = id;
@@ -16,4 +18,6 @@ public sealed class DownloadMigrationOptions(
     public bool Overwrite { get; } = overwrite;
     public TimeSpan MedianFirstRetryDelay { get; } =
         medianFirstRetryDelay ?? TimeSpan.FromMinutes(1);
+    public Action<string>? OnTemporaryFileCreated { get; } = onTemporaryFileCreated;
+    public Action<long, long?>? OnDownloadProgress { get; } = onDownloadProgress;
 }